Extract credentials setup
